collected tests in one file
