+++ server/__test__/index.test.js
@@ -3,7 +3,7 @@ const app = require('../app');
 const dbBuild = require('../database/config/build');
 const connection = require('../database/config/connection');
 const getUserByEmail = require('../database/queries/getEmail');
-const { getNews } = require('../database/queries');
+const { getNews, getFoodCategory } = require('../database/queries');
 
 describe('authentication', () => {
   beforeEach(() => dbBuild());
@@ -175,8 +175,41 @@ describe('authentication', () => {
 
   test('should get news from the news table', async () => {
     const { rows } = await getNews();
-    return expect(rows[5].content).toEqual(
+    const allContent = rows.map((row) => row.content);
+    return expect(allContent).toContain(
       'Researchers simulated a tailgating situation with a small group of overweight but healthy men and examined the impact of eating and drinking on their livers using blood tests and a liver scan.',
     );
+  });
+
+  describe('health news', () => {
+    it('router returns 200 after sending the news', async () => {
+      const { status } = await request(app)
+        .get('/api/v1/healthnews/')
+        .expect(200);
+      expect(status).toBe(200);
+    });
+  });
+
+  describe('getFoodCategory query', () => {
+    test('should get Food in Category from the userFoodRelation table', async () => {
+      const { rows } = await getFoodCategory();
+      return expect(rows).toEqual();
+    });
+  });
+
+  describe('food category', () => {
+    it('router returns 401 if there is no user logged', async () => {
+      const { status } = await request(app)
+        .get('/api/v1/user/0/category/1/food')
+        .expect(401);
+      expect(status).toBe(401);
+    });
+
+    it('router returns 200 and sends all foods in the table', async () => {
+      const { status } = await request(app)
+        .get('/api/v1/user/4/category/1/food')
+        .expect(200);
+      expect(status).toBe(200);
+    });
   });
 });
